requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
watchdog==3.0.0
openai==1.3.0
google-generativeai==0.3.0
requests==2.31.0
